Seperate keycloak from user-service
