+++ internal/handlers/users_v1_handler.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/redhatinsights/mbop/internal/config"
 	"github.com/redhatinsights/mbop/internal/service/keycloak"
+	keycloak_user_service "github.com/redhatinsights/mbop/internal/service/keycloak-user-service"
 	"github.com/redhatinsights/mbop/internal/service/ocm"
 )
 
@@ -75,25 +76,32 @@ func UsersV1Handler(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 
-		client, err := keycloak.NewKeyCloakClient()
-		if err != nil {
-			do400(w, err.Error())
-			return
-		}
-
-		err = client.InitKeycloakConnection()
+		keycloakClient := keycloak.NewKeyCloakClient()
+		err = keycloakClient.InitKeycloakConnection()
 		if err != nil {
 			do500(w, "Can't build keycloak connection: "+err.Error())
 			return
 		}
 
-		token, err := client.GetAccessToken()
+		token, err := keycloakClient.GetAccessToken()
 		if err != nil {
 			do500(w, "Can't fetch keycloak token: "+err.Error())
 			return
 		}
 
-		u, err := client.GetUsers(token, usernames, q)
+		userServiceClient, err := keycloak_user_service.NewKeyCloakUserServiceClient()
+		if err != nil {
+			do500(w, "Can't build keycloak user service client: "+err.Error())
+			return
+		}
+
+		err = userServiceClient.InitKeycloakUserServiceConnection()
+		if err != nil {
+			do500(w, "Can't build keycloak user service connection: "+err.Error())
+			return
+		}
+
+		u, err := userServiceClient.GetUsers(token, usernames, q)
 		if err != nil {
 			do500(w, "Cant Retrieve Keycloak Accounts: "+err.Error())
 			return
