+++ internal/service/keycloak/keycloak.go
@@ -6,72 +6,26 @@ import (
 	"io/ioutil"
 	"net/http"
 	"net/url"
-	"strconv"
 	"strings"
 	"time"
 
 	"github.com/redhatinsights/mbop/internal/config"
-	l "github.com/redhatinsights/mbop/internal/logger"
 	"github.com/redhatinsights/mbop/internal/models"
 )
 
-type Client struct {
+type KeycloakClient struct {
 	client *http.Client
 }
 
-func (keyCloak *Client) InitKeycloakConnection() error {
-	keyCloak.client = &http.Client{
+func (keycloak *KeycloakClient) InitKeycloakConnection() error {
+	keycloak.client = &http.Client{
 		Timeout: time.Duration(config.Get().KeyCloakTimeout * int64(time.Second)),
 	}
 
 	return nil
 }
 
-func (keyCloak *Client) GetUsers(token string, u models.UserBody, q models.UserV1Query) (models.Users, error) {
-	users := models.Users{Users: []models.User{}}
-	url, err := createV1RequestURL(u, q)
-	if err != nil {
-		return users, err
-	}
-
-	body, err := keyCloak.sendKeycloakGetRequest(url, token)
-	if err != nil {
-		l.Log.Error(err, "/v3/users error sending request")
-		return users, err
-	}
-
-	unmarshaledResponse := models.KeycloakResponses{}
-	err = json.Unmarshal(body, &unmarshaledResponse)
-	if err != nil {
-		return users, err
-	}
-
-	return keycloakResponseToUsers(unmarshaledResponse.Users), err
-}
-
-func (keyCloak *Client) GetAccountV3Users(orgID string, token string, q models.UserV3Query) (models.Users, error) {
-	users := models.Users{Users: []models.User{}}
-	url, err := createV3UsersRequestURL(orgID, q)
-	if err != nil {
-		return users, err
-	}
-
-	body, err := keyCloak.sendKeycloakGetRequest(url, token)
-	if err != nil {
-		l.Log.Error(err, "/v3/users error sending request")
-		return users, err
-	}
-
-	unmarshaledResponse := models.KeycloakResponses{}
-	err = json.Unmarshal(body, &unmarshaledResponse)
-	if err != nil {
-		return users, err
-	}
-
-	return keycloakResponseToUsers(unmarshaledResponse.Users), nil
-}
-
-func (keyCloak *Client) GetAccessToken() (string, error) {
+func (keycloak *KeycloakClient) GetAccessToken() (string, error) {
 	token := models.KeycloakTokenObject{}
 	url, err := createTokenURL()
 	if err != nil {
@@ -100,34 +54,6 @@ func (keyCloak *Client) GetAccessToken() (string, error) {
 	return token.AccessToken, nil
 }
 
-func (keyCloak *Client) sendKeycloakGetRequest(url *url.URL, token string) ([]byte, error) {
-	var responseBody []byte
-
-	req, err := http.NewRequest(http.MethodGet, url.String(), nil)
-	if err != nil {
-		return responseBody, err
-	}
-
-	req.Header.Set("Authorization", "Bearer "+token)
-
-	resp, err := keyCloak.client.Do(req)
-	if err != nil {
-		l.Log.Error(err, "error fetching keycloak response")
-		return responseBody, err
-	}
-
-	responseBody, err = ioutil.ReadAll(resp.Body)
-	if err != nil {
-		l.Log.Error(err, "error reading keycloak response body")
-		return responseBody, err
-	}
-
-	// Close response body
-	defer resp.Body.Close()
-
-	return responseBody, nil
-}
-
 func createEncodedTokenBody() *strings.Reader {
 	data := url.Values{}
 	data.Set("username", config.Get().KeyCloakTokenUsername)
@@ -139,90 +65,10 @@ func createEncodedTokenBody() *strings.Reader {
 }
 
 func createTokenURL() (*url.URL, error) {
-	url, err := url.Parse(fmt.Sprintf("%s://%s%s/token", config.Get().KeyCloakScheme, config.Get().KeyCloakHost, config.Get().KeyCloakPort))
+	url, err := url.Parse(fmt.Sprintf("%s%s", config.Get().KeyCloakTokenURL, config.Get().KeyCloakTokenPath))
 	if err != nil {
 		return nil, fmt.Errorf("error creating keycloak token url: %s", err)
 	}
 
 	return url, err
-}
-
-// MAKE response to users function to massage data back to regular format
-func createV1RequestURL(usernames models.UserBody, q models.UserV1Query) (*url.URL, error) {
-	url, err := url.Parse(fmt.Sprintf("%s://%s%s/users?limit=100", config.Get().KeyCloakScheme, config.Get().KeyCloakHost, config.Get().KeyCloakPort))
-	if err != nil {
-		return nil, fmt.Errorf("error creating (keycloak) /users url: %s", err)
-	}
-
-	queryParams := url.Query()
-
-	if q.QueryBy != "" {
-		queryParams.Add("order", q.QueryBy)
-	}
-
-	if q.SortOrder != "" {
-		queryParams.Add("direction", q.SortOrder)
-	}
-
-	queryParams.Add("usernames", createUsernamesQuery(usernames.Users))
-
-	url.RawQuery = queryParams.Encode()
-	return url, err
-}
-
-func createV3UsersRequestURL(orgID string, q models.UserV3Query) (*url.URL, error) {
-	url, err := url.Parse(fmt.Sprintf("%s://%s:%s/users", config.Get().KeyCloakScheme, config.Get().KeyCloakHost, config.Get().KeyCloakPort))
-	if err != nil {
-		return nil, fmt.Errorf("error creating (keycloak) /v3/users url: %s", err)
-	}
-	queryParams := url.Query()
-
-	if q.SortOrder != "" {
-		queryParams.Add("direction", q.SortOrder)
-	}
-
-	queryParams.Add("org_id", orgID)
-	queryParams.Add("limit", strconv.Itoa(q.Limit))
-	queryParams.Add("offset", strconv.Itoa(q.Offset))
-
-	url.RawQuery = queryParams.Encode()
-
-	return url, err
-}
-
-func createUsernamesQuery(usernames []string) string {
-	usernameQuery := ""
-
-	for _, username := range usernames {
-		if usernameQuery == "" {
-			usernameQuery += username
-		} else {
-			usernameQuery += fmt.Sprintf(",%s", username)
-		}
-	}
-
-	return usernameQuery
-}
-
-func keycloakResponseToUsers(r []models.KeycloakResponse) models.Users {
-	users := models.Users{Users: []models.User{}}
-
-	for _, response := range r {
-		users.AddUser(models.User{
-			Username:      response.Username,
-			ID:            response.ID,
-			Email:         response.Email,
-			FirstName:     response.FirstName,
-			LastName:      response.LastName,
-			AddressString: "",
-			IsActive:      true,
-			IsInternal:    response.IsInternal,
-			Locale:        "en_US",
-			OrgID:         response.OrgID,
-			DisplayName:   response.UserID,
-			Type:          response.Type,
-		})
-	}
-
-	return users
 }
