+++ internal/service/keycloak/keycloak_integration.go
@@ -1,0 +1,12 @@
+package keycloak
+
+type KeyCloak interface {
+	InitKeycloakConnection() error
+	GetAccessToken() (string, error)
+}
+
+func NewKeyCloakClient() KeyCloak {
+	client := &KeycloakClient{}
+
+	return client
+}
